Extract discussion round runner
